+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 - Add __eq__ method to Q to more easily test dynamically-built queries (#1506)
 - Added PlainToTsQuery function for postgres (#1347)
 - Allow field's default keyword to be async function (#1498)
+- Add support for queryset slicing. (#1341)
 
 Fixed
 ^^^^^